chore: type namespaces [MONET-1338]
