+++ packages/ecommerce-app-base/src/SkuPicker/ProductSelectionList/ProductSelectionList.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { css } from 'emotion';
-import { Product } from '../../interfaces';
+import { Product } from '../../types';
 import { ProductSelectionListItem } from './ProductSelectionListItem';
 import { OtherProductsCounter } from './OtherProductsCounter';
