+++ packages/ecommerce-app-base/src/__mocks__/productPreviews.ts
@@ -1,4 +1,4 @@
-import { Product } from '../interfaces';
+import { Product } from '../types';
 
 const productPreviews: Product[] = [
   {
