+++ packages/ecommerce-app-base/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './ui';
+export * from './config';
+export * from './product';
+export * from './integration';
