+++ packages/ecommerce-app-base/src/utils.ts
@@ -3,7 +3,7 @@
  * order that the SKUs were added to the field value.
  * @see https://gist.github.com/ecarter/1423674
  */
-import { Product } from './interfaces';
+import { Product } from './types';
 
 export const mapSort = <T extends Product, K extends keyof T>(
   array: T[],
